requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
openai==1.102.0
httpx==0.27.2
pydantic==2.8.2
tenacity==9.0.0
pytest==8.3.2
